Don't expose the OpenGL headers any place we don't need to.

+++ GutsyStorm/FoxTextureArray.h
@@ -7,12 +7,10 @@
 //
 
 #import <Cocoa/Cocoa.h>
-#import <OpenGL/gl.h>
-#import <OpenGL/OpenGL.h>
 
 @interface FoxTextureArray : NSObject
 
-- (nullable instancetype)initWithImagePath:(nonnull NSString *)path numTextures:(GLuint)numTextures;
+- (nullable instancetype)initWithImagePath:(nonnull NSString *)path numTextures:(NSUInteger)numTextures;
 - (void)bind;
 - (void)unbind;
